refactor: remove 'steps' folder
The steps folder doesn't add anything and creates a level which make importing
more difficult. It made sense for bundler 2 only.
