+++ packages/liferay-npm-bundler/src/adapt/angular-cli/templates/adapt-rt.js.ejs
