+++ packages/liferay-npm-bundler/src/adapt/angular-cli/tweakAttachmentToDOM.ts
