+++ packages/liferay-npm-bundler/src/adapt/bundler-project/index.ts
@@ -14,9 +14,9 @@ import {
 } from 'liferay-js-toolkit-core';
 import path from 'path';
 
-import {bundlerWebpackDir, project} from '../../../globals';
-import * as log from '../../../util/log';
-import Renderer from '../../../util/renderer';
+import {bundlerWebpackDir, project} from '../../globals';
+import * as log from '../../util/log';
+import Renderer from '../../util/renderer';
 import replaceWebpackJsonp from './replaceWebpackJsonp';
 import writeExportModules from './write-export-modules';
