+++ packages/liferay-npm-bundler/src/adapt/bundler-project/replaceWebpackJsonp.ts
