+++ packages/liferay-npm-bundler/src/adapt/create-react-app/templates/adapt-rt.js.ejs
