+++ packages/liferay-npm-bundler/src/adapt/create-react-app/tweakAttachmentToDOM.ts
