+++ packages/liferay-npm-bundler/src/adapt/templates/export-module.ejs
