+++ packages/liferay-npm-bundler/src/adapt/templates/index.js.ejs
