+++ packages/liferay-npm-bundler/src/adapt/templates/manifest-module.ejs
